+++ src/natcap/invest/urban_flood_risk_mitigation.py
@@ -632,7 +632,7 @@ def _build_affected_vector(
     infrastructure_layer = infrastructure_vector.GetLayer()
 
     infrastructure_srs = infrastructure_layer.GetSpatialRef()
-    infrastructure_to_target = osr.CoordinateTransformation(
+    infrastructure_to_target = utils.create_coordinate_transformer(
         infrastructure_srs, target_srs)
 
     infrastructure_layer_defn = infrastructure_layer.GetLayerDefn()